requirements.txt
numpy
pandas
sklearn
flask
nltk
newspaper3k
feedparser
python-dotenv
bs4
lxml
selenium
flask_mail
spacy
transformers
sentencepiece
torch